// supabase/functions/_shared/detection/types.ts
export type ContentType =
  | 'academic'
  | 'research'
  | 'blog'
  | 'seo'
  | 'news'
  | 'business'
  | 'email'
  | 'job'
  | 'legal'
  | 'technical'
  | 'creative'
  | 'social'
  | 'product'
  | 'student'
  | 'auto';